+++ calibre-plugin/prefs.py
@@ -33,6 +33,7 @@ prefs.defaults = {
     'ao3_password': '',
     'generate_covers': True,
     'set_calibre_cover': True,
+    'context_menu_placed': False,
     'koreader_path': '.adds/koreader',
     'include_prereleases': False,
 }